Update API DescribeOasAnomalySQLList: update response parameters Body.Data.$.AvgLogicalReads' format has changed.

+++ oceanbasepro-20190901/src/main/java/com/aliyun/oceanbasepro20190901/models/DescribeOasSQLHistoryListResponseBody.java
@@ -48,7 +48,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>9978.75</p>
          */
         @NameInMap("AvgAffectedRows")
-        public Long avgAffectedRows;
+        public Double avgAffectedRows;
 
         /**
          * <p>Average Application event wait time (in milliseconds) during the client waiting period.</p>
@@ -66,7 +66,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgBlockCacheHit")
-        public Long avgBlockCacheHit;
+        public Double avgBlockCacheHit;
 
         /**
          * <p>Average BlockIndexCache hit count during the execution period.</p>
@@ -75,7 +75,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgBlockIndexCacheHit")
-        public Long avgBlockIndexCacheHit;
+        public Double avgBlockIndexCacheHit;
 
         /**
          * <p>Average BloomFilterCache hit count during the execution period.</p>
@@ -84,7 +84,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgBloomFilterCacheHit")
-        public Long avgBloomFilterCacheHit;
+        public Double avgBloomFilterCacheHit;
 
         /**
          * <p>Average Concurrency event wait time (in milliseconds) during the execution period.</p>
@@ -129,7 +129,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgDiskReads")
-        public Long avgDiskReads;
+        public Double avgDiskReads;
 
         /**
          * <p>Average response time (in milliseconds) during the execution period.</p>
@@ -183,7 +183,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgLogicalReads")
-        public Long avgLogicalReads;
+        public Double avgLogicalReads;
 
         /**
          * <p>Average Memstore read rows during the execution period.</p>
@@ -192,7 +192,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgMemstoreReadRows")
-        public Long avgMemstoreReadRows;
+        public Double avgMemstoreReadRows;
 
         /**
          * <p>Average network transmission time (in milliseconds) during the execution period.</p>
@@ -237,7 +237,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgReturnRows")
-        public Long avgReturnRows;
+        public Double avgReturnRows;
 
         /**
          * <p>Average RowCache hit count during the execution period.</p>
@@ -246,7 +246,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgRowCacheHit")
-        public Long avgRowCacheHit;
+        public Double avgRowCacheHit;
 
         /**
          * <p>Average number of RPC sent during the execution.</p>
@@ -255,7 +255,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>8.0</p>
          */
         @NameInMap("AvgRpcCount")
-        public Long avgRpcCount;
+        public Double avgRpcCount;
 
         /**
          * <p>Average Schedule event wait time (in milliseconds) during the scheduling period.</p>
@@ -273,7 +273,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0.0</p>
          */
         @NameInMap("AvgSsstoreReadRows")
-        public Long avgSsstoreReadRows;
+        public Double avgSsstoreReadRows;
 
         /**
          * <p>Average number of threads used of the SQL during the execution period.</p>
@@ -498,7 +498,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode4012Count")
-        public Double retCode4012Count;
+        public Long retCode4012Count;
 
         /**
          * <p>Number of occurrences of code 4013 during the execution period.</p>
@@ -507,7 +507,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode4013Count")
-        public Double retCode4013Count;
+        public Long retCode4013Count;
 
         /**
          * <p>Number of occurrences of code 5001 during the execution period.</p>
@@ -516,7 +516,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode5001Count")
-        public Double retCode5001Count;
+        public Long retCode5001Count;
 
         /**
          * <p>Number of occurrences of code 5024 during the execution period.</p>
@@ -525,7 +525,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode5024Count")
-        public Double retCode5024Count;
+        public Long retCode5024Count;
 
         /**
          * <p>Number of occurrences of code 5167 during the execution period.</p>
@@ -534,7 +534,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode5167Count")
-        public Double retCode5167Count;
+        public Long retCode5167Count;
 
         /**
          * <p>Number of occurrences of code 5217 during the execution period.</p>
@@ -543,7 +543,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode5217Count")
-        public Double retCode5217Count;
+        public Long retCode5217Count;
 
         /**
          * <p>Number of occurrences of code 6002 during the execution period.</p>
@@ -552,7 +552,7 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
          * <p>0</p>
          */
         @NameInMap("RetCode6002Count")
-        public Double retCode6002Count;
+        public Long retCode6002Count;
 
         /**
          * <p>Total number of retries during the execution period.</p>
@@ -676,11 +676,11 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return TeaModel.build(map, self);
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgAffectedRows(Long avgAffectedRows) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgAffectedRows(Double avgAffectedRows) {
             this.avgAffectedRows = avgAffectedRows;
             return this;
         }
-        public Long getAvgAffectedRows() {
+        public Double getAvgAffectedRows() {
             return this.avgAffectedRows;
         }
 
@@ -692,27 +692,27 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return this.avgApplicationWaitTime;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgBlockCacheHit(Long avgBlockCacheHit) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgBlockCacheHit(Double avgBlockCacheHit) {
             this.avgBlockCacheHit = avgBlockCacheHit;
             return this;
         }
-        public Long getAvgBlockCacheHit() {
+        public Double getAvgBlockCacheHit() {
             return this.avgBlockCacheHit;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgBlockIndexCacheHit(Long avgBlockIndexCacheHit) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgBlockIndexCacheHit(Double avgBlockIndexCacheHit) {
             this.avgBlockIndexCacheHit = avgBlockIndexCacheHit;
             return this;
         }
-        public Long getAvgBlockIndexCacheHit() {
+        public Double getAvgBlockIndexCacheHit() {
             return this.avgBlockIndexCacheHit;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgBloomFilterCacheHit(Long avgBloomFilterCacheHit) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgBloomFilterCacheHit(Double avgBloomFilterCacheHit) {
             this.avgBloomFilterCacheHit = avgBloomFilterCacheHit;
             return this;
         }
-        public Long getAvgBloomFilterCacheHit() {
+        public Double getAvgBloomFilterCacheHit() {
             return this.avgBloomFilterCacheHit;
         }
 
@@ -748,11 +748,11 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return this.avgDecodeTime;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgDiskReads(Long avgDiskReads) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgDiskReads(Double avgDiskReads) {
             this.avgDiskReads = avgDiskReads;
             return this;
         }
-        public Long getAvgDiskReads() {
+        public Double getAvgDiskReads() {
             return this.avgDiskReads;
         }
 
@@ -796,19 +796,19 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return this.avgGetPlanTime;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgLogicalReads(Long avgLogicalReads) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgLogicalReads(Double avgLogicalReads) {
             this.avgLogicalReads = avgLogicalReads;
             return this;
         }
-        public Long getAvgLogicalReads() {
+        public Double getAvgLogicalReads() {
             return this.avgLogicalReads;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgMemstoreReadRows(Long avgMemstoreReadRows) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgMemstoreReadRows(Double avgMemstoreReadRows) {
             this.avgMemstoreReadRows = avgMemstoreReadRows;
             return this;
         }
-        public Long getAvgMemstoreReadRows() {
+        public Double getAvgMemstoreReadRows() {
             return this.avgMemstoreReadRows;
         }
 
@@ -844,27 +844,27 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return this.avgQueueTime;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgReturnRows(Long avgReturnRows) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgReturnRows(Double avgReturnRows) {
             this.avgReturnRows = avgReturnRows;
             return this;
         }
-        public Long getAvgReturnRows() {
+        public Double getAvgReturnRows() {
             return this.avgReturnRows;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgRowCacheHit(Long avgRowCacheHit) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgRowCacheHit(Double avgRowCacheHit) {
             this.avgRowCacheHit = avgRowCacheHit;
             return this;
         }
-        public Long getAvgRowCacheHit() {
+        public Double getAvgRowCacheHit() {
             return this.avgRowCacheHit;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgRpcCount(Long avgRpcCount) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgRpcCount(Double avgRpcCount) {
             this.avgRpcCount = avgRpcCount;
             return this;
         }
-        public Long getAvgRpcCount() {
+        public Double getAvgRpcCount() {
             return this.avgRpcCount;
         }
 
@@ -876,11 +876,11 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return this.avgScheduleTime;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setAvgSsstoreReadRows(Long avgSsstoreReadRows) {
+        public DescribeOasSQLHistoryListResponseBodyData setAvgSsstoreReadRows(Double avgSsstoreReadRows) {
             this.avgSsstoreReadRows = avgSsstoreReadRows;
             return this;
         }
-        public Long getAvgSsstoreReadRows() {
+        public Double getAvgSsstoreReadRows() {
             return this.avgSsstoreReadRows;
         }
 
@@ -1076,59 +1076,59 @@ public class DescribeOasSQLHistoryListResponseBody extends TeaModel {
             return this.remotePlans;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode4012Count(Double retCode4012Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode4012Count(Long retCode4012Count) {
             this.retCode4012Count = retCode4012Count;
             return this;
         }
-        public Double getRetCode4012Count() {
+        public Long getRetCode4012Count() {
             return this.retCode4012Count;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode4013Count(Double retCode4013Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode4013Count(Long retCode4013Count) {
             this.retCode4013Count = retCode4013Count;
             return this;
         }
-        public Double getRetCode4013Count() {
+        public Long getRetCode4013Count() {
             return this.retCode4013Count;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode5001Count(Double retCode5001Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode5001Count(Long retCode5001Count) {
             this.retCode5001Count = retCode5001Count;
             return this;
         }
-        public Double getRetCode5001Count() {
+        public Long getRetCode5001Count() {
             return this.retCode5001Count;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode5024Count(Double retCode5024Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode5024Count(Long retCode5024Count) {
             this.retCode5024Count = retCode5024Count;
             return this;
         }
-        public Double getRetCode5024Count() {
+        public Long getRetCode5024Count() {
             return this.retCode5024Count;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode5167Count(Double retCode5167Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode5167Count(Long retCode5167Count) {
             this.retCode5167Count = retCode5167Count;
             return this;
         }
-        public Double getRetCode5167Count() {
+        public Long getRetCode5167Count() {
             return this.retCode5167Count;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode5217Count(Double retCode5217Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode5217Count(Long retCode5217Count) {
             this.retCode5217Count = retCode5217Count;
             return this;
         }
-        public Double getRetCode5217Count() {
+        public Long getRetCode5217Count() {
             return this.retCode5217Count;
         }
 
-        public DescribeOasSQLHistoryListResponseBodyData setRetCode6002Count(Double retCode6002Count) {
+        public DescribeOasSQLHistoryListResponseBodyData setRetCode6002Count(Long retCode6002Count) {
             this.retCode6002Count = retCode6002Count;
             return this;
         }
-        public Double getRetCode6002Count() {
+        public Long getRetCode6002Count() {
             return this.retCode6002Count;
         }
